feat: [WD-19698] Custom ISO permission check.
- Custom ISO - can delete
- Codebase refactor.
