+++ src/pages/storage/UploadCustomIso.tsx
@@ -10,7 +10,6 @@ import {
   Input,
   Notification,
 } from "@canonical/react-components";
-import { createIsoStorageVolume } from "api/storage-pools";
 import { useCurrentProject } from "context/useCurrentProject";
 import Loader from "components/Loader";
 import ProgressBar from "components/ProgressBar";
@@ -23,6 +22,7 @@ import type { AxiosError } from "axios";
 import type { LxdSyncResponse } from "types/apiResponse";
 import { isValidISOAlias, sanitizeISOAlias } from "util/customISO";
 import classnames from "classnames";
+import { createIsoStorageVolume } from "api/storage-volumes";
 
 interface Props {
   onFinish: (name: string, pool: string) => void;
