+++ src/pages/storage/actions/DeleteStorageVolumeBtn.tsx
@@ -1,7 +1,6 @@
 import type { FC } from "react";
 import { useState } from "react";
 import type { LxdStorageVolume } from "types/storage";
-import { deleteStorageVolume } from "api/storage-pools";
 import { useQueryClient } from "@tanstack/react-query";
 import { queryKeys } from "util/queryKeys";
 import {
@@ -9,6 +8,8 @@ import {
   Icon,
   useNotify,
 } from "@canonical/react-components";
+import { useStorageVolumeEntitlements } from "util/entitlements/storage-volumes";
+import { deleteStorageVolume } from "api/storage-volumes";
 
 interface Props {
   volume: LxdStorageVolume;
@@ -32,6 +33,7 @@ const DeleteStorageVolumeBtn: FC<Props> = ({
   const notify = useNotify();
   const [isLoading, setLoading] = useState(false);
   const queryClient = useQueryClient();
+  const { canDeleteVolume } = useStorageVolumeEntitlements();
 
   const getDisabledReason = () => {
     if (volume.name.includes("/")) {
@@ -104,8 +106,12 @@ const DeleteStorageVolumeBtn: FC<Props> = ({
       className={classname}
       shiftClickEnabled
       showShiftClickHint
-      disabled={Boolean(disabledReason)}
-      onHoverText={disabledReason}
+      disabled={!canDeleteVolume(volume) || Boolean(disabledReason)}
+      onHoverText={
+        canDeleteVolume(volume)
+          ? disabledReason
+          : "You do not have permission to delete this volume."
+      }
     >
       {hasIcon && <Icon name="delete" />}
       {label && <span>{label}</span>}
