+++ src/pages/storage/actions/snapshots/VolumeConfigureSnapshotModal.tsx
@@ -12,10 +12,10 @@ import type { LxdStorageVolume } from "types/storage";
 import type { StorageVolumeFormValues } from "pages/storage/forms/StorageVolumeForm";
 import { volumeFormToPayload } from "pages/storage/forms/StorageVolumeForm";
 import { getStorageVolumeEditValues } from "util/storageVolumeEdit";
-import { updateStorageVolume } from "api/storage-pools";
 import StorageVolumeFormSnapshots from "pages/storage/forms/StorageVolumeFormSnapshots";
 import { useToastNotification } from "context/toastNotificationProvider";
 import ResourceLink from "components/ResourceLink";
+import { updateStorageVolume } from "api/storage-volumes";
 
 interface Props {
   volume: LxdStorageVolume;
