+++ src/util/entitlements/storage-volumes.tsx
@@ -1,0 +1,18 @@
+import { useAuth } from "context/auth";
+import { hasEntitlement } from "./helpers";
+import type { LxdStorageVolume } from "types/storage";
+
+export const useStorageVolumeEntitlements = () => {
+  const { isFineGrained } = useAuth();
+
+  const canDeleteVolume = (volume?: LxdStorageVolume) =>
+    hasEntitlement(isFineGrained, "can_delete", volume?.access_entitlements);
+
+  const canEditVolume = (volume?: LxdStorageVolume) =>
+    hasEntitlement(isFineGrained, "can_edit", volume?.access_entitlements);
+
+  return {
+    canDeleteVolume,
+    canEditVolume,
+  };
+};
